Extracted AST functionality to new class.
